config: remove top-level theme
